Cleaned up home page and implement responsivity
Removed a number of not needed files as well as cleaned up a number of elements that were not needed in the home page file.

Added different UI elements to the NavBar and created appropriate class names to be displayed or hidden based on media queries. Updated CSS files to reflect this in both NavBar .jsx/css and Footer.jsx/css files in the components folder.

Also removed ReadME file and Readme screenshots